pyproject.toml
[tool.poetry]
name = "cryscript"
version = "0.1.0"
description = ""
authors = ["Daniel Kim <kdanielive@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ccxt = "^4.4.33"
ipython = "^8.29.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
